tee: optee: Allow to freeze while waiting for call_queue
When the system is going to hibernate or suspend and the tasks of
the other entries in the call queue are frozen wait_for_completion
on the call_queue might get stuck.

Change wait to interruptible and add try_to_freeze in order to
allow that the waiting task is frozen as well.

Signed-off-by: Christoph Gellner <cgellner@de.adit-jv.com>
Signed-off-by: Dung Nguyen <dung.nguyen.zy@renesas.com>
Signed-off-by: Nhi Nguyen <nhi.nguyen.kx@renesas.com>