BSP rebase: Drop two commits for simple-bridge not applicable upstream
For commit e42ca15fced1 ("drm: bridge: simple-bridge: Add preferred
resolution option"), the initial preferred resolution is not a property
of the board and should not be reported in DTS.

For commit 8605b0f29f33 ("drm: bridge: simple-bridge: Add connector
status setting option"), the kernel shouldn't cheat when it can't detect
the monitor. Userspace should be able to cope with connectors have an
unknown state. If there's a valid reason why userspace can't handle
these connectors correctly, the exact reason and use case needs to be
reported, to be discussed with upstream.

Signed-off-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>