BSP rebase: Ignore CSI-2 for V4H
The CSI-2 registers for R8A779G0 V4H is not documented in the Hardware
User's Manual [1]. There is not much we can do about rebaseing this work
without documentation.

Move the effected commits to the ignore file to really highlight to the
BSP team that this is not handled by the upstream team.

1. R19UH0172EJ0070 Rev.0.70 Jan. 2023
   r19uh0186ej0070-r-carv4h.pdf

Signed-off-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>