BSP rebase: Ignore GMSL2 and FPD-Link III commits
GMSL2 and FPD-Link III support is a project that needs to be scheduled
on its own. It is now easier to do so as we we have landed the V4L2
streams API upstream, but the amount of work remains too large to be
handled within the scope of generic BSP upporting.

Mark the related commits as ignored, and create a task file to cover
this work.

Signed-off-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>