BSP rebase: Mark V4H video capture patches as handled
Mark all video capture patches for V4H as handled. This mark is special
as it resurrects the CSI-2 patches from the ignores file.

When working with the CSI-2 driver to test the other capture related
components I could remove all the really nasty undocumented bits from
the CSI-2 driver, leaving us with more or less the documentation status
we have for CSI-2 on Gen3, which is acceptable.

The only thing not documented that would be nice is the lane swapping
register. This is now hard-coded, but once we have documentation for it
it should be trivial to map this to the data-lanes and clock-lanes
properties. I find this trade-off acceptable as we then can move forward
with hopefully a upstream driver for V4H CSI-2.

In addition the R-Car capture patches C-PHY support of the MAX96712 have
been added to the upstream staging driver. The staging driver is
different from the one used in the BSP, but any effort to upstream work
around MAX96712 should be aimed at the upstream staging driver. So we
might just as well get C-PHY support out of the way. In addition this
allows the V4H capture pipeline to be tested as the MAX96712 staging
driver generates test patterns.

Signed-off-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>