BSP rebase: Ignore THS interrupt-mode outside trip-points
This is not how the thermal core expect things to work. The core always
pulls the driver for the temperature reading reported to user-space, and
the (optional) interrupts are used to handle the trip-points. The idea
is that an interrupt shall only fire once we move past a trip point so
that the system can take the action associated with that trip-point
ASAP.

If there is no interrupt support on the system the trip-point will still
be trigged by the core when it pulls the temperature.

Signed-off-by: Niklas S将ﾓderlund <niklas.soderlund+renesas@ragnatech.se>