LTS: Ignore power settings for debug pins
The patch in question modifies the power source for settings for 4 pins
on Sparrow Hawk; GP_4_0, GP_5_0, GP_6_0 and GP_7_0. These pins are used
as followed from the schematics.

    pin         name            connected too
    GP_4_0	TSN0_MDIO	U8001 (Marvell 88Q2110)
    GP_5_0	AVB2_AVTP_PPS   SW47
    GP_6_0	AVB1_MDIO	U8007 (Marvell 88Q2110)
    GP_7_0	AVB0_AVTP_PPS	GPLED1

The two pins that map to AVB{0,2}_AVTP_PPS are connected to the debug
LEDs and push switch, do not effect operation of the Ethernet. The other
two pins map to the MDIO bus of AVB0 and TSN0. This could effect
operation of networking but these interfaces have been proved to work
without this change.

The original commits messaging indicating this change would touch on all
AVB0-3 and TSN0, which it clearly do not. Conclusion is this is
something done to debug or test non-network related things that uses
pins in GPIO mode. As the commit message do not describe what was
tested, nor the intent move the patch to the ignore list.

Signed-off-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>